feat: refactored styles, seperated TG group component into a sep. file
