+++ src/pages/interfaces.ts
@@ -1,0 +1,2 @@
+export type StatusType = "error" | "success";
+export type TelegramStatusType = "error" | "info" | "success";
